chore: more unit tests
