+++ packages/frontend/src/components/dialogs/manage-transaction/helpers.test.ts
@@ -1,12 +1,18 @@
-import { TRANSACTION_TYPES } from '@bt/shared/types';
+import { OUT_OF_WALLET_ACCOUNT_MOCK, VERBOSE_PAYMENT_TYPES } from '@/common/const';
+import { PAYMENT_TYPES, TRANSACTION_TRANSFER_NATURE, TRANSACTION_TYPES } from '@bt/shared/types';
 import {
+  ACCOUNTS,
+  USER_CATEGORIES,
   buildOutOfWalletTransaction,
   buildSystemExpenseTransaction,
   buildSystemIncomeTransaction,
   buildSystemTransferExpenseTransaction,
+  buildSystemTransferOppositeTransaction,
+  getUah2Account,
+  getUahAccount,
 } from '@tests/mocks';
 
-import { getDestinationAmount, getFormTypeFromTransaction, getTxTypeFromFormType } from './helpers';
+import { getDestinationAmount, getFormTypeFromTransaction, getTxTypeFromFormType, prepopulateForm } from './helpers';
 import { FORM_TYPES } from './types';
 
 describe('components/modals/modify-record/helpers', () => {
@@ -50,6 +56,207 @@ describe('components/modals/modify-record/helpers', () => {
       [FORM_TYPES.transfer, TRANSACTION_TYPES.expense],
     ])('%s to be %s', (value, expected) => {
       expect(getTxTypeFromFormType(value)).toBe(expected);
+    });
+  });
+
+  describe('prepopulateForm', () => {
+    const accountsRecord = ACCOUNTS.reduce(
+      (acc, account) => {
+        acc[account.id] = account;
+        return acc;
+      },
+      {} as Record<number, (typeof ACCOUNTS)[0]>,
+    );
+
+    const categoriesRecord = USER_CATEGORIES.reduce(
+      (acc, category) => {
+        acc[category.id] = category;
+        return acc;
+      },
+      {} as Record<number, (typeof USER_CATEGORIES)[0]>,
+    );
+
+    it('returns undefined when transaction is undefined', () => {
+      const result = prepopulateForm({
+        transaction: undefined,
+        oppositeTransaction: undefined,
+        accounts: accountsRecord,
+        categories: categoriesRecord,
+      });
+
+      expect(result).toBeUndefined();
+    });
+
+    it('populates form for basic income transaction', () => {
+      const transaction = buildSystemIncomeTransaction();
+
+      const result = prepopulateForm({
+        transaction,
+        oppositeTransaction: undefined,
+        accounts: accountsRecord,
+        categories: categoriesRecord,
+      });
+
+      expect(result).toMatchObject({
+        type: FORM_TYPES.income,
+        amount: transaction.amount,
+        account: accountsRecord[transaction.accountId],
+        category: categoriesRecord[transaction.categoryId],
+        note: transaction.note,
+        refundedByTxs: undefined,
+        refundsTx: undefined,
+      });
+      expect(result.time).toBeInstanceOf(Date);
+      expect(result.paymentType.value).toBe(transaction.paymentType);
+    });
+
+    it('populates form for basic expense transaction', () => {
+      const transaction = buildSystemExpenseTransaction();
+
+      const result = prepopulateForm({
+        transaction,
+        oppositeTransaction: undefined,
+        accounts: accountsRecord,
+        categories: categoriesRecord,
+      });
+
+      expect(result).toMatchObject({
+        type: FORM_TYPES.expense,
+        amount: transaction.amount,
+        account: accountsRecord[transaction.accountId],
+        category: categoriesRecord[transaction.categoryId],
+      });
+    });
+
+    it('populates form for transfer transaction with opposite transaction', () => {
+      const sourceAccount = getUahAccount();
+      const destinationAccount = getUah2Account();
+
+      const transaction = buildSystemTransferExpenseTransaction({
+        accountId: sourceAccount.id,
+        amount: 1000,
+      });
+      const oppositeTransaction = buildSystemTransferOppositeTransaction({
+        accountId: destinationAccount.id,
+        amount: 1500,
+        transferId: transaction.transferId,
+      });
+
+      const result = prepopulateForm({
+        transaction,
+        oppositeTransaction,
+        accounts: accountsRecord,
+        categories: categoriesRecord,
+      });
+
+      expect(result).toMatchObject({
+        type: FORM_TYPES.transfer,
+        amount: transaction.amount,
+        account: sourceAccount,
+        toAccount: destinationAccount,
+        targetAmount: oppositeTransaction.amount,
+      });
+    });
+
+    it('populates form for out-of-wallet income transaction (external → account)', () => {
+      const destinationAccount = getUahAccount();
+
+      const transaction = buildSystemIncomeTransaction({
+        accountId: destinationAccount.id,
+        amount: 500,
+        transferNature: TRANSACTION_TRANSFER_NATURE.transfer_out_wallet,
+        transactionType: TRANSACTION_TYPES.income,
+      });
+
+      const result = prepopulateForm({
+        transaction,
+        oppositeTransaction: undefined,
+        accounts: accountsRecord,
+        categories: categoriesRecord,
+      });
+
+      expect(result).toMatchObject({
+        type: FORM_TYPES.transfer,
+        account: OUT_OF_WALLET_ACCOUNT_MOCK,
+        toAccount: destinationAccount,
+        targetAmount: transaction.amount,
+      });
+      // amount should not be set for income out-of-wallet
+      expect(result.amount).toBeUndefined();
+    });
+
+    it('populates form for out-of-wallet expense transaction (account → external)', () => {
+      const sourceAccount = getUahAccount();
+
+      const transaction = buildSystemExpenseTransaction({
+        accountId: sourceAccount.id,
+        amount: 750,
+        transferNature: TRANSACTION_TRANSFER_NATURE.transfer_out_wallet,
+        transactionType: TRANSACTION_TYPES.expense,
+      });
+
+      const result = prepopulateForm({
+        transaction,
+        oppositeTransaction: undefined,
+        accounts: accountsRecord,
+        categories: categoriesRecord,
+      });
+
+      expect(result).toMatchObject({
+        type: FORM_TYPES.transfer,
+        account: sourceAccount,
+        toAccount: OUT_OF_WALLET_ACCOUNT_MOCK,
+        amount: transaction.amount,
+      });
+      // targetAmount should not be set for expense out-of-wallet
+      expect(result.targetAmount).toBeUndefined();
+    });
+
+    it('preserves transaction note', () => {
+      const transaction = buildSystemExpenseTransaction({
+        note: 'Test note for transaction',
+      });
+
+      const result = prepopulateForm({
+        transaction,
+        oppositeTransaction: undefined,
+        accounts: accountsRecord,
+        categories: categoriesRecord,
+      });
+
+      expect(result.note).toBe('Test note for transaction');
+    });
+
+    it('maps payment type correctly', () => {
+      const transaction = buildSystemExpenseTransaction({
+        paymentType: PAYMENT_TYPES.cash,
+      });
+
+      const result = prepopulateForm({
+        transaction,
+        oppositeTransaction: undefined,
+        accounts: accountsRecord,
+        categories: categoriesRecord,
+      });
+
+      expect(result.paymentType).toEqual(VERBOSE_PAYMENT_TYPES.find((p) => p.value === PAYMENT_TYPES.cash));
+    });
+
+    it('converts transaction time to Date object', () => {
+      const txTime = new Date('2024-06-15T10:30:00Z');
+      const transaction = buildSystemExpenseTransaction({
+        time: txTime,
+      });
+
+      const result = prepopulateForm({
+        transaction,
+        oppositeTransaction: undefined,
+        accounts: accountsRecord,
+        categories: categoriesRecord,
+      });
+
+      expect(result.time).toBeInstanceOf(Date);
+      expect(result.time.getTime()).toBe(txTime.getTime());
     });
   });
 });
